refactoring the food code
