+++ src/components/UI/Molecules/BiotipoSelector.tsx
@@ -1,4 +1,4 @@
-import { BIOTIPO_DESCRIPTIONS } from '@/utils/BiotiposDescriptios';
+import { BIOTIPO_DESCRIPTIONS } from '@/components/statics/BiotiposDescriptios';
 import { BiotipoBox } from '../Atoms/BiotipoBox';
 import { BiotipoSelectorProps } from '@/interfaces/BiotipoSelectorProps';
