+++ src/components/UI/Molecules/FoodOption.tsx
@@ -1,10 +1,8 @@
 import { FoodOptionProps } from '@/interfaces/FoodOptionProps';
 import { FoodBox } from '../Atoms/FoodBox';
-import { GiChickenOven, GiManualJuicer, GiMilkCarton, GiSlicedBread } from 'react-icons/gi';
-import { FaAppleAlt, FaCarrot, FaEgg, FaFish } from 'react-icons/fa';
-import { TbMeat } from 'react-icons/tb';
-import { CiWheat } from 'react-icons/ci';
-import { PiPlant } from 'react-icons/pi';
+import { foodItems } from '@/components/statics/foodNames';
+import { isFoodDisabled } from '@/functions/food-disabled';
+import { getFoodIcon } from '@/functions/get-food-icon';
 
 export const FoodOption: React.FC<FoodOptionProps> = ({ foodSelections, setFoodSelections, selectedDiet }) => {
 
@@ -16,49 +14,6 @@ export const FoodOption: React.FC<FoodOptionProps> = ({ foodSelections, setFoodS
     });
   };;
 
-  const getIcon = (food: string) => {
-    switch (food) {
-      case 'Aves':
-        return <GiChickenOven className="w-8 h-8" />;
-      case 'Peixe':
-        return <FaFish className="w-8 h-8" />;
-      case 'Carne':
-        return <TbMeat className="w-8 h-8" />;
-      case 'Leite':
-        return <GiMilkCarton className="w-8 h-8" />;
-      case 'Fruta':
-        return <FaAppleAlt className="w-8 h-8" />;
-      case 'Legumes':
-        return <FaCarrot className="w-8 h-8" />;
-      case 'Ovo':
-        return <FaEgg className="w-8 h-8" />;
-      case 'Cereais':
-        return <CiWheat className="w-8 h-8" />;
-      case 'Pães':
-        return <GiSlicedBread className="w-8 h-8" />;
-      case 'Suco':
-        return <GiManualJuicer className="w-8 h-8" />;
-      case 'Vegetais':
-        return <PiPlant className="w-8 h-8" />;
-      default:
-        return null;
-    }
-  };
-
-  const foodItems = [
-    "Aves", "Peixe", "Carne", "Leite", "Fruta", "Legumes", "Ovo", "Cereais", "Pães", "Suco", "Vegetais"
-  ];
-
-  const isDisabled = (food: string) => {
-    if (selectedDiet === "vegana" && ["Aves", "Peixe", "Carne", "Leite", "Ovo"].includes(food)) {
-      return true;
-    }
-    if (selectedDiet === "vegetariana" && ["Aves", "Peixe", "Carne"].includes(food)) {
-      return true;
-    }
-    return false;
-  };
-
   return (
     <div className="flex flex-wrap justify-center gap-4 mt-5">
       {foodItems.map(food => (
@@ -66,9 +21,9 @@ export const FoodOption: React.FC<FoodOptionProps> = ({ foodSelections, setFoodS
           key={food}
           title={food}
           isSelected={!!foodSelections[food]}
-          icon={getIcon(food)}
+          icon={getFoodIcon(food)}
           onSelect={() => handleSelectFood(food)}
-          isDisabled={isDisabled(food)}
+          isDisabled={isFoodDisabled(selectedDiet, food)}
         />
       ))}
     </div>
