+++ src/components/statics/BiotiposDescriptios.ts
