+++ src/components/statics/foodNames.ts
@@ -1,0 +1,14 @@
+export const foodItems = [
+    "Aves", 
+    "Peixe", 
+    "Carne", 
+    "Leite", 
+    "Fruta", 
+    "Legumes", 
+    "Ovo", 
+    "Cereais", 
+    "Pães", 
+    "Suco", 
+    "Vegetais"
+];
+
