+++ src/components/statics/steps.ts
